+++ eg_next/src/app/page.tsx
@@ -1,6 +1,7 @@
 import Image from 'next/image';
 import { Inter } from 'next/font/google';
 import TestComp from './TestComp';
+import Sidebar from '@/shared/components/newcomps/Sidebar';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -9,6 +10,7 @@ export default function Home() {
     <main className="flex min-h-screen flex-col items-center justify-between p-24">
       <h1> hi next 2 </h1>
       <TestComp></TestComp>
+      <Sidebar />
     </main>
   );
 }
