+++ src/boxModel/other.ts
@@ -1,16 +1,19 @@
+import { createCommandBasic, createCommandCustom, createCommandSize } from "./createCommand";
+
 const otherCommandsArray: any = [];
 
-otherCommandsArray.push({
-  name: "objectCover",
-  func: () => "object-fit: cover",
-});
-otherCommandsArray.push({
-  name: "relative",
-  func: () => "position: relative",
-});
-otherCommandsArray.push({
-  name: "absolute",
-  func: () => "position: absolute",
-});
+
+otherCommandsArray.push(
+  createCommandBasic('objectCover', 'object-fit: cover'),
+  createCommandBasic('relative', 'position: relative'),
+  createCommandBasic('absolute', 'position: absolute'),
+
+  createCommandSize('top', 'top'),
+  createCommandSize('bottom', 'bottom'),
+  createCommandSize('left', 'left'),
+  createCommandSize('right', 'right'),
+
+  createCommandCustom('zIndex', (x: any) => `z-index: ${x}`),
+);
 
 export { otherCommandsArray };
